Move custom Layout classes to separate file
